refactor(bridge): update label of transfer status
